# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SiliconSentinel orchestrator",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/target/profile": {
      "get": {
        "summary": "Profile",
        "operationId": "profile_api_v1_target_profile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/target/status": {
      "get": {
        "summary": "Status",
        "operationId": "status_api_v1_target_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/uart/status": {
      "get": {
        "summary": "Uart Status",
        "operationId": "uart_status_api_v1_uart_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildIdentity": {
        "properties": {
          "component": {
            "type": "string",
            "enum": [
              "edge",
              "orchestrator"
            ],
            "title": "Component"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "dirty": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dirty"
          },
          "source_fingerprint": {
            "type": "string",
            "title": "Source Fingerprint"
          },
          "build_id": {
            "type": "string",
            "title": "Build Id"
          },
          "process_started_at": {
            "type": "string",
            "title": "Process Started At"
          },
          "pid": {
            "type": "integer",
            "title": "Pid"
          },
          "python": {
            "type": "string",
            "title": "Python"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "component",
          "version",
          "source_fingerprint",
          "build_id",
          "process_started_at",
          "pid",
          "python"
        ],
        "title": "BuildIdentity",
        "description": "Non-secret identity of the code a process actually loaded; see contracts/build.py."
      },
...
      "DebuggerStatus": {
        "properties": {
          "applicable": {
            "type": "boolean",
            "title": "Applicable"
          },
          "reachable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reachable"
          },
          "version_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Ok"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "expected_version_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version Prefix"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "applicable"
        ],
        "title": "DebuggerStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Region": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start": {
            "type": "integer",
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Start"
          },
          "end": {
            "type": "integer",
            "maximum": 4294967296.0,
            "minimum": 1.0,
            "title": "End"
          },
          "kind": {
            "type": "string",
            "enum": [
              "ram",
              "rom"
            ],
            "title": "Kind"
          },
          "address_space": {
            "type": "string",
            "enum": [
              "virtual",
              "physical"
            ],
            "title": "Address Space",
            "default": "virtual"
          },
          "executable": {
            "type": "boolean",
            "title": "Executable",
            "default": false
          },
          "instruction_mode": {
            "type": "string",
            "enum": [
              "arm",
              "thumb"
            ],
            "title": "Instruction Mode",
            "default": "arm"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "start",
          "end",
          "kind"
        ],
        "title": "Region"
      },
...
      "TargetProfile": {
        "properties": {
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "architecture": {
            "type": "string",
            "const": "arm",
            "title": "Architecture",
            "default": "arm"
          },
          "endianness": {
            "type": "string",
            "enum": [
              "little",
              "big"
            ],
            "title": "Endianness",
            "default": "little"
          },
          "address_width": {
            "type": "integer",
            "const": 32,
            "title": "Address Width",
            "default": 32
          },
          "address_spaces": {
            "items": {
              "type": "string",
              "enum": [
                "virtual",
                "physical"
              ]
            },
            "type": "array",
            "title": "Address Spaces",
            "default": [
              "virtual"
            ]
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/Region"
            },
            "type": "array",
            "title": "Regions"
          },
          "registers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Registers",
            "default": [
              "pc",
              "lr",
              "sp",
              "cpsr"
            ]
          },
          "capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "read",
                "registers",
                "snapshot",
                "halt",
                "resume",
                "step",
                "write"
              ]
            },
            "type": "array",
            "title": "Capabilities",
            "default": [
              "read",
              "registers",
              "snapshot"
            ]
          },
          "provenance": {
            "type": "string",
            "title": "Provenance"
          },
          "verified_for_live": {
            "type": "boolean",
            "title": "Verified For Live",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_id",
          "regions",
          "provenance"
        ],
        "title": "TargetProfile"
      },
      "TargetStatus": {
        "properties": {
          "communication": {
            "type": "string",
            "enum": [
              "ok",
              "failed",
              "unknown"
            ],
            "title": "Communication"
          },
          "execution_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "communication"
        ],
        "title": "TargetStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "EdgeStatus": {
        "properties": {
          "schema_name": {
            "type": "string",
            "const": "jtagent.edge.status",
            "title": "Schema Name"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At"
          },
          "application": {
            "type": "string",
            "const": "ok",
            "title": "Application"
          },
          "build": {
            "$ref": "#/components/schemas/BuildIdentity"
          },
          "source_drift": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Drift"
          },
          "debugger": {
            "$ref": "#/components/schemas/DebuggerStatus"
          },
          "target": {
            "$ref": "#/components/schemas/TargetStatus"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "target_backend": {
            "type": "string",
            "enum": [
              "mock",
              "replay",
              "openocd"
            ],
            "title": "Target Backend"
          },
          "generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation"
          },
          "capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "read",
                "registers",
                "snapshot",
                "halt",
                "resume",
                "step",
                "write"
              ]
            },
            "type": "array",
            "title": "Capabilities"
          },
          "recovery_required": {
            "type": "boolean",
            "title": "Recovery Required"
          },
          "snapshots_armed": {
            "type": "boolean",
            "title": "Snapshots Armed"
          },
          "retains_raw_locally": {
            "type": "boolean",
            "title": "Retains Raw Locally"
          },
          "read_prerequisite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read Prerequisite"
          },
          "live_operations": {
            "$ref": "#/components/schemas/LiveGate"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_name",
          "schema_version",
          "observed_at",
          "application",
          "build",
          "debugger",
          "target",
          "target_backend",
          "generation",
          "capabilities",
          "recovery_required",
          "snapshots_armed",
          "retains_raw_locally",
          "live_operations"
        ],
        "title": "EdgeStatus",
        "description": "Edge status v2. The unversioned flat shape (state, target_backend, ...) is schema v1."
      },
      "LiveGate": {
        "properties": {
          "allowed": {
            "type": "boolean",
            "title": "Allowed"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "allowed",
          "blockers"
        ],
        "title": "LiveGate"
      },
...
    }
  }
}